adapted seeding to mongoDB Atlas + fixed cloudinary whitelisting

+++ seeds/seedUsers.js
@@ -17,9 +17,12 @@ const User = require('../models/users');
 
 module.exports = async function () {
 
-    await User.deleteMany({});
-    console.log("Users reset done before seeding");
-
+    const usersExist = await User.find();
+    if (usersExist) {
+        await User.deleteMany({});
+        console.log("Users reset done before seeding");
+    }
+    
     const users = ["Helena Thomson", "Fran Martinez", "Mike Smith", "Paula Colomé"];
     await Promise.all(users.map( async (user) => {
         let u = new User({
